lion/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
boto3==1.28.50